Add files via upload
The client (Arduino) is generate token from the result of hashed timestamp, salt (phrase), and distance (data sensor) because I'm using ultrasonic sensor. After that, the server is trying to verify the token from the client before. If the token is same before the 2 platforms, it will appear "Token Verified" and show some data. But if the token is not same, it will appear "Token Verification Failed" and won't show any data.